+++ lapce-rpc/src/plugin.rs
@@ -24,14 +24,14 @@ pub struct PluginConfiguration {
 }
 
 #[derive(Deserialize, Clone, Debug, Serialize)]
-#[serde(rename_all = "kebab-case")]
 pub struct VoltInfo {
     pub name: String,
     pub version: String,
     pub display_name: String,
     pub author: String,
     pub description: String,
-    pub meta: String,
+    pub repository: Option<String>,
+    pub wasm: bool,
 }
 
 impl VoltInfo {
@@ -83,7 +83,8 @@ impl VoltMetadata {
             display_name: self.display_name.clone(),
             author: self.author.clone(),
             description: self.description.clone(),
-            meta: "".to_string(),
+            repository: self.repository.clone(),
+            wasm: self.wasm.is_some(),
         }
     }
 }